requirements.txt
streamlit
pandas
pyyaml
plotly
openpyxl
altair
watchdog
